+++ packages/client/src/api.ts
@@ -50,7 +50,6 @@ class ApiClient {
       });
     }
 
-    // @ts-ignore
     axiosRetry(this.axiosClient, {
       retries: 3,
       retryCondition: this.canRetryRequest,