+++ packages/client/src/clients/feed/feed.ts
@@ -1,3 +1,4 @@
+import { GenericData } from "@knocklabs/types";
 import EventEmitter from "eventemitter2";
 import { Channel } from "phoenix";
 import { StoreApi } from "zustand";
@@ -42,9 +43,9 @@ class Feed {
   private defaultOptions: FeedClientOptions;
   private broadcastChannel!: BroadcastChannel | null;
   private disconnectTimer: ReturnType<typeof setTimeout> | null = null;
-  private hasSubscribedToRealTimeUpdates: Boolean = false;
+  private hasSubscribedToRealTimeUpdates: boolean = false;
   private visibilityChangeHandler: () => void = () => {};
-  private visibilityChangeListenerConnected: Boolean = false;
+  private visibilityChangeListenerConnected: boolean = false;
 
   // The raw store instance, used for binding in React and other environments
   public store: StoreApi<FeedStoreState>;
@@ -659,7 +660,7 @@ class Feed {
     }
   }
 
-  private broadcastOverChannel(type: string, payload: any) {
+  private broadcastOverChannel(type: string, payload: GenericData) {
     // The broadcastChannel may not be available in non-browser environments
     if (!this.broadcastChannel) {
       return;