+++ packages/client/src/clients/feed/interfaces.ts
@@ -94,7 +94,7 @@ export interface FeedMetadata {
   unseen_count: number;
 }
 
-export interface FeedResponse<T = any> {
+export interface FeedResponse<T = GenericData> {
   entries: FeedItem<T>[];
   meta: FeedMetadata;
   page_info: PageInfo;