+++ packages/client/src/clients/feed/store.ts
@@ -1,10 +1,17 @@
-import create from "zustand/vanilla";
+import { GenericData } from "@knocklabs/types";
+import zustand from "zustand/vanilla";
 
 import { NetworkStatus } from "../../networkStatus";
 
 import { FeedItem } from "./interfaces";
 import { FeedStoreState } from "./types";
 import { deduplicateItems, sortItems } from "./utils";
+
+// Get the correct Zustand function. Caused by some issues in v3 exports
+// https://github.com/pmndrs/zustand/issues/334
+const create: typeof zustand =
+  // @ts-expect-error workaround for issue above
+  typeof zustand === "function" ? zustand : zustand.default;
 
 function processItems(items: FeedItem[]) {
   const deduped = deduplicateItems(items);
@@ -53,7 +60,7 @@ export default function createStore() {
       set((state) => {
         // We resort the list on set, so concating everything is fine (if a bit suboptimal)
         const items = options.shouldAppend
-          ? processItems(state.items.concat(entries))
+          ? processItems(state.items.concat(entries as FeedItem<GenericData>[]))
           : entries;
 
         return {