+++ packages/client/src/clients/messages/interfaces.ts
@@ -1,3 +1,5 @@
+import { GenericData } from "@knocklabs/types";
+
 import { RecipientRef } from "../..";
 
 export type MessageDeliveryStatus =
@@ -20,7 +22,7 @@ export type MessageEngagementStatus =
   | "interacted"
   | "archived";
 
-export interface Message<T = any> {
+export interface Message<T = GenericData> {
   id: string;
   channel_id: string;
   recipient: RecipientRef;