+++ packages/client/src/clients/objects/index.ts
@@ -1,3 +1,5 @@
+import { GenericData } from "@knocklabs/types";
+
 import { ApiResponse } from "../../api";
 import { ChannelData } from "../../interfaces";
 import Knock from "../../knock";
@@ -12,7 +14,7 @@ type SetChannelDataInput = {
   objectId: string;
   collection: string;
   channelId: string;
-  data: any;
+  data: GenericData;
 };
 
 class ObjectClient {
@@ -21,7 +23,7 @@ class ObjectClient {
   constructor(instance: Knock) {
     this.instance = instance;
   }
-  async getChannelData<T = any>({
+  async getChannelData<T = GenericData>({
     collection,
     objectId,
     channelId,