+++ packages/react/src/index.ts
@@ -1,4 +1,5 @@
 import "./theme.css";
+
 export * from "./modules/core";
 export * from "./modules/feed";
 export * from "./modules/slack";